feat/basic-landing page. more to come
